logout fix

+++ app/Jobs/LogOutUsers.php
@@ -39,9 +39,11 @@ class LogOutUsers implements ShouldQueue
         foreach ($sessions as $session) {
             $timePassed = $currentTimestamp - $session->last_activity;
 
-            if ($timePassed >= $hour) {
-                  $session->user()->addLogout();
-                  $session->delete();
+            if($session->user()) {
+              if ($timePassed >= $hour) {
+                    $session->user()->addLogout();
+                    $session->delete();
+              }
             }
         }
     }
